requirements.txt
# Основные зависимости
requests # Библиотека для HTTP-запросов

# Инструменты для тестирования
pytest # Фреймворк для тестирования
selenium # Инструмент для автоматизации браузера
webdriver-manager # Утилита для управления веб-драйверами
allure-pytest # Плагин для интеграции Allure с pytest

# Линтеры и инструменты для анализа кода
flake8 # Линтер для проверки стиля кода